Refactor map classes for replay
